fix(admin-directory): moved clerk provider, changed admin directory
moved clerk provider to admin/dashboard layout only, changed (admin) to admin and moved auth pages
