+++ .env.example
@@ -19,8 +19,8 @@ XATA_HTTPS_DATABASE_URL=
 
 ##clerk
 NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY=
-NEXT_PUBLIC_CLERK_SIGN_IN_URL=/sign-in
-NEXT_PUBLIC_CLERK_SIGN_UP_URL=/sign-up
+NEXT_PUBLIC_CLERK_SIGN_IN_URL=/admin/sign-in
+NEXT_PUBLIC_CLERK_SIGN_UP_URL=/admin/sign-up
 NEXT_PUBLIC_CLERK_AFTER_SIGN_IN_URL=/dashboard
 NEXT_PUBLIC_CLERK_AFTER_SIGN_UP_URL=/dashboard
