+++ src/app/admin/dashboard/layout.tsx
@@ -1,5 +1,5 @@
 //info: must import env types file so validation is run
-import { auth } from '@clerk/nextjs'
+import { ClerkProvider, auth } from '@clerk/nextjs'
 import { isAdminUser } from '@/utils/auth'
 import { redirect } from 'next/navigation'
 import AdminNav from '@/components/dashboard/AdminNav'
@@ -26,13 +26,15 @@ export default async function Home({
   }
 
   return (
-    <main className="mx-auto space-y-12 px-4 pb-10">
-      <section className="mx-auto grid gap-8 space-y-12 px-4  pb-10 lg:grid-cols-3 lg:space-y-0">
-        <div className="mx-auto lg:col-span-1">
-          <AdminNav />
-        </div>
-        <div className="mx-auto lg:col-span-2">{children}</div>
-      </section>
-    </main>
+    <ClerkProvider>
+      <main className="mx-auto space-y-12 px-4 pb-10">
+        <section className="mx-auto grid gap-8 space-y-12 px-4  pb-10 lg:grid-cols-3 lg:space-y-0">
+          <div className="mx-auto lg:col-span-1">
+            <AdminNav />
+          </div>
+          <div className="mx-auto lg:col-span-2">{children}</div>
+        </section>
+      </main>
+    </ClerkProvider>
   )
 }
