+++ src/app/layout.tsx
@@ -1,7 +1,5 @@
 import type { Metadata } from 'next'
 import './globals.css'
-import { ClerkProvider } from '@clerk/nextjs'
-import Nav from '@/components/nav/Nav'
 import Footer from '@/components/Footer'
 import GlobalSearch from '@/components/search/GlobalSearch'
 import { SearchProvider } from '@/components/search/SearchContext'
@@ -9,6 +7,7 @@ import { Toaster } from 'react-hot-toast'
 import PlausibleProvider from 'next-plausible'
 
 import { Raleway } from 'next/font/google'
+import Nav from '@/components/nav/Nav'
 const raleway = Raleway({
   subsets: ['latin'],
   display: 'swap',
@@ -38,19 +37,17 @@ export default function RootLayout({
         <PlausibleProvider domain="dealsfordevs.com" />
       </head>
       <body className="">
-        <ClerkProvider>
-          <div className="flex min-h-screen flex-col justify-between bg-gray-900">
-            <div className="mx-auto w-full max-w-screen-2xl px-2 pt-5 md:px-24 xl:pt-20">
-              <SearchProvider>
-                <GlobalSearch />
-                <Nav />
-              </SearchProvider>
-              <div>{children}</div>
-            </div>
-            <Footer />
+        <div className="flex min-h-screen flex-col justify-between bg-gray-900">
+          <div className="mx-auto w-full max-w-screen-2xl px-2 pt-5 md:px-24 xl:pt-20">
+            <SearchProvider>
+              <GlobalSearch />
+              <Nav />
+            </SearchProvider>
+            <div>{children}</div>
           </div>
-          <Toaster position="top-center" toastOptions={{ duration: 3000 }} />
-        </ClerkProvider>
+          <Footer />
+        </div>
+        <Toaster position="top-center" toastOptions={{ duration: 3000 }} />
       </body>
     </html>
   )
