+++ src/components/dashboard/AdminNav.tsx
@@ -1,5 +1,4 @@
 'use client'
-import { useState } from 'react'
 import Link from 'next/link'
 
 export default function AdminNav() {
